project/airflow/requirements.txt
apache-airflow-providers-google==10.19.0
google-cloud-storage==2.16.0
pandas
pyarrow==15.0.2